requirements.txt
Django==1.11
django-filter==1.1.0
djangorestframework==3.8.1
nose==1.3.7
pytz==2018.4
